+++ packages/react-native-sdk/src/utils/push/rxSubjects.ts
@@ -61,3 +61,13 @@ export const voipCallkeepCallOnForegroundMap$ = new BehaviorSubject<
 export const voipCallkeepAcceptedCallOnNativeDialerMap$ = new BehaviorSubject<
   CallkeepMap | undefined
 >(undefined);
+
+type UnsubscribeCallback = () => void;
+
+/**
+ * This rxjs subject is used to store the unsubscribe callbacks (if any) of the push notification processing
+ * Note: it should be used to clear it when app processes push notification from foreground
+ */
+export const pushUnsubscriptionCallbacks$ = new BehaviorSubject<
+  UnsubscribeCallback[] | undefined
+>(undefined);