+++ sample-apps/react/react-dogfood/pages/join/[callId].tsx
@@ -33,7 +33,6 @@ import type {
   CreateJwtTokenRequest,
   CreateJwtTokenResponse,
 } from '../api/auth/create-token';
-import { BackgroundFiltersProvider } from '../../components/BackgroundFilters';
 
 const basePath = process.env.NEXT_PUBLIC_BASE_PATH || '';
 
@@ -210,20 +209,7 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
       >
         <StreamCall call={call}>
           <TourProvider>
-            <BackgroundFiltersProvider
-              isBlurringEnabled={true}
-              backgroundImages={[
-                `${basePath}/backgrounds/amsterdam-1.jpg`,
-                `${basePath}/backgrounds/amsterdam-2.jpg`,
-                `${basePath}/backgrounds/boulder-1.jpg`,
-                `${basePath}/backgrounds/boulder-2.jpg`,
-                `${basePath}/backgrounds/gradient-1.jpg`,
-                `${basePath}/backgrounds/gradient-2.jpg`,
-                `${basePath}/backgrounds/gradient-3.jpg`,
-              ]}
-            >
-              <MeetingUI chatClient={chatClient} />
-            </BackgroundFiltersProvider>
+            <MeetingUI chatClient={chatClient} />
           </TourProvider>
         </StreamCall>
       </StreamVideo>