+++ sample-apps/react/react-video-demo/src/components/ChatMessage/index.ts
@@ -1,1 +1,0 @@
-export { ChatMessage as default } from './ChatMessage';