+++ sample-apps/react/react-video-demo/src/components/ChatSendButton/index.ts
@@ -1,1 +1,0 @@
-export { ChatSendButton as default } from './ChatSendButton';