+++ sample-apps/react/react-video-demo/src/components/Header/Header.data.ts
@@ -1,6 +1,0 @@
-import { HeaderProps } from './Header';
-
-export const KichinSink: HeaderProps = {
-  callId: '12321dDA2354da',
-  isCallActive: false,
-};