+++ sample-apps/react/react-video-demo/src/components/LatencyMapPopup/LatencyMapPopup.data.ts
@@ -1,7 +1,0 @@
-import { Props } from './LatencyMapPopup';
-
-export const KichinSink: Props = {
-  abbriviation: 'BOS',
-  city: 'Boston',
-  countryCode: 'USA',
-};